==> src/R/06_match_norcca_names.R <==
library(tidyverse)
library(worrms)
library(snakecase)
library(rvest)

# Define which origin countries to keep
nordic_countries <- c("Sweden", "Denmark", "Norway", 
                      "Finland", "Iceland", "Faeroe Islands", 
                      "Poland", "Lithuania", "Latvia", 
                      "Estonia", "Germany", "United Kingdom", 
                      "England", "Ireland", "Scotland")

# Read NORCCA taxa lists
norcca_worms <- read_tsv("data_in/norcca_names_matched.txt",
                         col_types = cols()) # List of taxa that has been matched with the WoRMS web match interface
norcca_strains <- read_tsv("data_in/norcca_strains.txt", # Generated by NORCCA compiler
                           col_types = cols())
taxa_worms <- read_tsv("data_out/content/taxa.txt",
                       col_types = cols())

# Fix column names and select relevant information
norcca_worms <- norcca_worms %>%
  select(scientific_name, status, AphiaID, AphiaID_accepted)

# Get all aphia_id
aphia_id <- unique(norcca_worms$AphiaID)

# Remove NA
aphia_id <- aphia_id[!is.na(aphia_id)]

# Load stored file if running from cache
if(file.exists("cache/norcca_cache.rda")) {
  load(file = "cache/norcca_cache.rda")
} else { 
  norcca_updated <- data.frame()
}

# Skip cached items
aphia_id <- aphia_id[!aphia_id %in% norcca_updated$AphiaID]

# Loop for each AphiaID to get taxonomic records
if (length(aphia_id) > 0) {
  for(i in 1:length(aphia_id)) {
    tryCatch({
      record <- wm_record(aphia_id[i])
      
      norcca_updated <- rbind(norcca_updated, record)
      
      save(norcca_updated, file = "cache/norcca_cache.rda")
    }, error=function(e){
      cat("Error occurred in AphiaID", aphia_id[i], ":", conditionMessage(e), "\n")
    }
    )
  }
}

# Update AphiaID if id is unaccepted
norcca_worms <- norcca_updated %>%
  select(AphiaID, valid_AphiaID, status, valid_name) %>%
  right_join(select(norcca_worms,
                    -status,
                    -AphiaID_accepted), 
             by = "AphiaID") %>%
  mutate(taxon_id = ifelse(status == "unaccepted", valid_AphiaID, AphiaID)) %>%
  mutate(taxon_id = ifelse(status == "deleted", valid_AphiaID, taxon_id)) %>%
  select(-valid_AphiaID, -status, -AphiaID)

# Find taxa without AphiaID
missing_id <- norcca_strains %>%
  filter(!Species %in% norcca_worms$scientific_name) %>%
  mutate(scientific_name = gsub(" sp.", "", Species)) %>%
  mutate(scientific_name = str_trim(scientific_name))

# Create empty df
identified_taxa <- data.frame()

# Loop for each name to get AphiaID
for(i in 1:nrow(missing_id)) {
  tryCatch({
  missing_i <- missing_id[i,]
  
  missing_i$AphiaID = wm_name2id(missing_i$scientific_name)
  
  identified_taxa <- rbind(identified_taxa, missing_i)
  }, error=function(e){
    cat("Error occurred in iteration", i, ":", conditionMessage(e), "\n")
  })
}

# Find all AphiaIDs
aphia_id <- unique(identified_taxa$AphiaID)

# Extract records from WoRMS based on AphiaID
missing_records <- data.frame()

# Loop for each AphiaID to get taxonomic records
for(i in 1:length(aphia_id)) {
  tryCatch({
  record <- wm_record(aphia_id[i])
  
  missing_records <- rbind(missing_records, record)
  
  cat('Getting missing record', i, 'of', length(aphia_id),'\n')
  }, error=function(e){
    cat("Error occurred in AphiaID", aphia_id[i], ":", conditionMessage(e), "\n")
  })
}

# Keep only necessary info
missing_info <- identified_taxa %>%
  left_join(missing_records, by = "AphiaID") %>%
  select(Species, valid_name, status, AphiaID, valid_AphiaID) %>%
  rename(scientific_name = Species) %>%
  mutate(taxon_id = ifelse(status == "unaccepted", valid_AphiaID, AphiaID)) %>%
  mutate(taxon_id = ifelse(status == "deleted", valid_AphiaID, taxon_id)) %>%
  select(-valid_AphiaID, -AphiaID, -status) %>%
  distinct()

# Combine manually matched names with the missing info
norcca_worms <- bind_rows(norcca_worms, missing_info) %>%
  filter(!is.na(taxon_id))

# Join tables
norcca_combined <- norcca_strains %>%
  mutate(scientific_name = Species) %>%
  left_join(norcca_worms, by = "scientific_name") %>%
  mutate(scientific_name = gsub(" sp.", "", scientific_name)) %>%
  select(-Division, -Class) %>%
  # rename(taxon_id = used_aphia_id) %>%
  filter(taxon_id %in% taxa_worms$taxon_id) %>%
  mutate(url_slug = gsub('.*strain/', '', `Strain Link`)) %>%
  relocate(taxon_id, scientific_name, valid_name)

# Make snakecase headers
names(norcca_combined) <- snakecase::to_snake_case(names(norcca_combined))

# Create an empty column
norcca_combined$nordic <- NA

# Load stored file if running from cache
if(file.exists("cache/norcca_nordic_cache.rda")) {
  load(file = "cache/norcca_nordic_cache.rda")
} else { 
  norcca_nordic <- data.frame()
}

norcca_combined_missing <- norcca_combined %>%
  filter(!strain_link %in% norcca_nordic$strain_link)

# Scrape origin information from strain page
if (nrow(norcca_combined_missing) > 0) {
  for (i in 1:nrow(norcca_combined_missing)) {
    # message("Scraping strain page ", i, " of ", nrow(norcca_combined_missing))
    
    # Try to read the HTML, handle errors with tryCatch
    html <- tryCatch({
      read_html(norcca_combined_missing$strain_link[i])
    }, error = function(e) {
      message("Error at strain page ", i, ": ", e$message)
      return(NA)  # Return NA on error
    })
    
    # If read_html fails, set nordic[i] to NA
    if (is.na(html)) {
      norcca_combined_missing$nordic[i] <- NA
    } else {
      # Extract and inspect the legends of all fieldsets
      fieldsets <- html %>% html_elements("fieldset")
      legends <- fieldsets %>% html_elements("legend") %>% html_text()

      # Adjusted filtering using trimws() to remove any leading/trailing spaces
      origin_fieldset <- fieldsets %>%
        keep(~ trimws(html_element(.x, "legend") %>% html_text()) == "ORIGIN") %>%
        html_element(".fieldset-wrapper")
      
      # Extract the text content of the 'Origin' section
      origin_text <- origin_fieldset %>%
        html_text(trim = TRUE)
      
      # Check if any nordic country is mentioned in the 'Origin' section
      norcca_combined_missing$nordic[i] <- any(sapply(nordic_countries, function(country) grepl(country, origin_text)))
    }
  }
}

# Bind with cached items
norcca_nordic <- rbind(norcca_nordic,
                       norcca_combined_missing) %>%
  filter(!is.na(nordic)) %>%
  distinct()

save(norcca_nordic, file = "cache/norcca_nordic_cache.rda")

# Remove strains that are not defined as nordic countries
norcca_nordic <- norcca_nordic %>%
  filter(nordic) %>%
  filter(taxon_id %in% taxa_worms$taxon_id) %>%
  select(-nordic)

# Store file
write_tsv(norcca_nordic, "data_out/content/facts_external_links_norcca.txt", na = "")
